Implementado o Delete de Course.

+++ Services/Course/CourseService.cs
@@ -56,6 +56,10 @@ public class CourseService : IService<Course>
 
     public void Delete(long id)
     {
+        Course course = GetById(id);
+        if (_auth.LoggedInUser.Id != course.OwnerId)
+            throw new AccessDeniedException("Você não tem permissão para deletar este curso.");
+
         _uow.CourseRepository.DeleteById(id);
     }
 }
